Add files via upload
The Bitcoin_private_key_cracker program is designed to generate private bitcoin keys, and then check for the relevance of the private key and check the balance of the wallet. Found wallets with a positive balance are saved to the good.txt file and bad wallets are saved to the Bad_Keys.txt file. The program works on Windows systems from 7 to 11 versions of Windows.